Made some minor changes to the layout of stat.cpp. I then noticed that I could safely store variables about stats in a uint16_t instead of just an unsigned. I had to update my max damage tests to handle the decreased range. This led to an improvement in the damage calculator, including several bug fixes related to damage overflow.

+++ src/stat.hpp
@@ -42,8 +42,8 @@ class Stat {
 			END
 		};
 
-		unsigned max;		// Max HP only
-		unsigned stat;		// Current HP or last calculated value for other stats
+		uint16_t max;		// Max HP only
+		uint16_t stat;		// Current HP or last calculated value for other stats
 		uint8_t base;
 		int8_t iv;			// 0 through 31
 		int8_t ev;			// 0 through 63
